fix routing paths with fewer parts than a pattern
When a path contains fewer parts than a pattern, `Router.action_for`
ignores the latter parts of the pattern. This can cause paths to be
routed incorrectly. For example, `/foo` can be routed to the pattern
`/foo/bar`.

Fix this by using the maximum of the path and pattern lengths,
and failing when the end of one is reached before the other.